+++ tools/utilities/pythonlibs/audio/record.py
@@ -19,8 +19,8 @@ import microphone
 def main(rate, channels, duration):
     sample_width = 2  # 16 bit
 
-    mic = microphone.Microphone()
-    mic.open(512, rate, channels)
+    mic = microphone.Microphone(auto_scale=False)
+    mic.open(1024, rate, channels)
 
     start = time.time()
 
@@ -32,7 +32,6 @@ def main(rate, channels, duration):
         writer.setframerate(rate)
         while time.time() < start + duration:
             buffer = mic.read()
-            buffer *= 32768  # unscale it back to int16
             int16 = buffer.astype(np.int16)
             writer.writeframes(int16)
 
@@ -44,5 +43,11 @@ if __name__ == "__main__":
     arg_parser.add_argument("--sample_rate", "-r", help="The sample rate for the recording", default=16000, type=int)
     arg_parser.add_argument("--channels", "-c", help="The number of channels to record", default=1, type=int)
     arg_parser.add_argument("--duration", "-d", help="The number of seconds to record", default=1, type=int)
+    arg_parser.add_argument("--list_devices", help="List available input devices", action="store_true")
+    arg_parser.add_argument("--input_device", "-i", help="Index of input device (see --list_devices)",
+                            default=1, type=int)
     args = arg_parser.parse_args()
-    main(args.sample_rate, args.channels, args.duration)
+    if args.list_devices:
+        microphone.list_devices()
+    else:
+        main(args.sample_rate, args.channels, args.duration)