+++ tools/utilities/pythonlibs/audio/training/make_training_list.py
@@ -15,13 +15,40 @@ import os
 import numpy as np
 
 
+def is_background_noise(filename):
+    return "_background_noise_" in filename
+
+
 def load_list_file(filename):
     with open(filename, "r") as fp:
         return [e.strip() for e in fp.readlines()]
 
 
-def is_background_noise(filename):
-    return "_background_noise_" in filename
+def save_list_file(data, filename):
+    with open(filename, "w") as f:
+        for line in data:
+            f.write(line)
+            f.write('\n')
+
+
+def clean_list_file(filename, bad_list):
+    if not os.path.isfile(filename):
+        print("Cannot find file: {}".format(filename))
+        return
+
+    testlist = load_list_file(filename)
+
+    count = 0
+    cleanlist = []
+    for line in testlist:
+        if line in bad_list:
+            count += 1
+        else:
+            cleanlist += [line]
+
+    if count > 0:
+        print("removed {} bad files from {}".format(count, filename))
+        save_list_file(cleanlist, filename)
 
 
 def shuffle(filename, output):
@@ -30,16 +57,13 @@ def shuffle(filename, output):
         print("Cannot find file: {}".format(filename))
         return
 
-    with open(filename, "r") as f:
-        testlist = f.readlines()
+    testlist = load_list_file(filename)
 
-    testlist = np.random.choice(testlist, len(testlist), replace=False)
-
-    with open(output, "w") as f:
-        f.writelines(testlist)
+    shuffled = np.random.choice(testlist, len(testlist), replace=False)
+    save_list_file(shuffled, output)
 
 
-def make_training_list(wav_files, max_files_per_directory):
+def make_training_list(wav_files, max_files_per_directory, bad_list=None):
     """
     Create a training list file given the directory where the wav files are organized into subdirectories,
     with one subdirectory per keyword to be recognized.  This training list will exclude any files
@@ -50,6 +74,15 @@ def make_training_list(wav_files, max_files_per_directory):
         print("wav_file directory not found")
         return
 
+    bad = []
+    if bad_list:
+        if not os.path.isfile(bad_list):
+            bad_list = os.path.join(wav_files, bad_list)
+            if not os.path.isfile(bad_list):
+                print("Bad list {} not found".format(bad_list))
+            else:
+                bad = load_list_file(bad_list)
+
     testing_list = os.path.join(wav_files, "testing_list.txt")
     if not os.path.isfile(testing_list):
         print("### error: testing list '{}' file not found".format(testing_list))
@@ -59,6 +92,12 @@ def make_training_list(wav_files, max_files_per_directory):
     if not os.path.isfile(validation_list):
         print("### error: validation list '{}' file not found".format(validation_list))
         return
+
+    if len(bad) > 0:
+        print("Cleaning: {}".format(testing_list))
+        clean_list_file(testing_list, bad)
+        print("Cleaning: {}".format(validation_list))
+        clean_list_file(validation_list, bad)
 
     print("Shuffling: {}".format(testing_list))
     shuffle(testing_list, testing_list)
@@ -87,6 +126,8 @@ def make_training_list(wav_files, max_files_per_directory):
             for n in files:
                 if os.path.splitext(n)[1] == ".wav":
                     entry = dir_name + "/" + n
+                    if entry in bad:
+                        continue
                     if entry in ignore_list:
                         skipped += 1
                     else:
@@ -98,7 +139,7 @@ def make_training_list(wav_files, max_files_per_directory):
                 max = len(file_list)
 
             # shuffle the training list also
-            file_list = np.random.choice(np.array(file_list), max)
+            file_list = np.random.choice(np.array(file_list), max, replace=False)
             print("choosing {} random files from {}".format(len(file_list), dir_name))
 
             for e in file_list:
@@ -124,6 +165,8 @@ up to the given maximum number of files per directory.")
                             help="Maximum number of files to include from each subdirectory (default: 5000)",
                             type=int, default=5000)
     arg_parser.add_argument("--wav_files", "-w", help="Directory containing the wav files to process", required=True)
+    arg_parser.add_argument("--bad_list", "-b",
+                            help="List of bad files to exclude from testing, training and validation lists")
     args = arg_parser.parse_args()
 
-    make_training_list(args.wav_files, args.max_files_per_directory)
+    make_training_list(args.wav_files, args.max_files_per_directory, args.bad_list)